oidc/requirements.txt
async-timeout==4.0.2
black==23.1.0
blacksheep==2.0a5
certifi==2022.12.7
cffi==1.15.1
chardet==5.0.0
click==8.1.3
cryptography==38.0.4
essentials==1.1.5
essentials-openapi==1.0.6
guardpost==1.0.1
h11==0.14.0
httptools==0.5.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
redis==4.5.1
rodi==2.0.2
ruff==0.0.253
six==1.16.0
uvicorn==0.20.0
websockets==10.4
